+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.9] - 2026-02-15
+## [0.8.11] - 2026-02-15
+
+### Added
+
+- Seconds and Milisecond input to `TimePicker`
+
+### Changed
+
+- `DateUtils.toInputString` to allow for different precision levels
+
+## [0.8.10] - 2026-02-15
 
 ### Fixed